changed doubles to float

+++ include/dataset_parser.h
@@ -5,5 +5,5 @@ int parse_file_rows(char* filename);
 
 int parse_file_columns(char* filename);
 
-void parse_file_data(char* filename, double* dataset, double* target_values, int num_columns, int num_rows);
+void parse_file_data(char* filename, float* dataset, float* target_values, int num_columns, int num_rows);
 #endif
